refactor main lib code into submodules
